# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /outreach/inventory/import:
    post:
      operationId: importVerifiedOutreachInventory
      tags: [outreach]
      summary: Import the bundled verified business inventory (Clerk admin only)
      responses:
        "201": { description: Import report, content: { application/json: { schema: { $ref: "#/components/schemas/InventoryImportReport" } } } }
        "400": { description: Invalid or unsafe source, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorMessage" } } } }
  /outreach/lane-config:
    get:
      operationId: getOutreachLaneConfig
      tags: [outreach]
      parameters:
        - { name: campaignKey, in: query, required: false, schema: { type: string } }
      responses:
        "200": { description: Persistent runtime lane configuration, content: { application/json: { schema: { $ref: "#/components/schemas/OutreachLaneConfig" } } } }
  /outreach/system-config:
    get:
      operationId: getOutreachSystemConfig
      tags: [outreach]
      summary: Read-only effective outreach schedule and policy
      responses:
        "200": { description: Effective authoritative configuration, content: { application/json: { schema: { $ref: "#/components/schemas/OutreachSystemConfig" } } } }
        "503": { description: Configuration unavailable }
  /outreach/one-time-override:
    post:
      operationId: createOutreachOneTimeOverride
      tags: [outreach]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutreachOneTimeOverrideInput"
      responses:
        "201": { description: Audited blocked/requested override, content: { application/json: { schema: { $ref: "#/components/schemas/OutreachOneTimeOverride" } } } }
        "400": { description: Invalid request, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorMessage" } } } }
...
components:
  schemas:
    InventoryImportReport:
      type: object
      required: [batchId, sourceRows, importedPendingQualification, excluded, namedImported, publicImported, namedEligible, publicEligible, exclusions]
      properties:
        batchId: { type: string }
        sourceRows: { type: number }
        imported: { type: number }
        importedPendingQualification: { type: number }
        excluded: { type: number }
        namedImported: { type: number }
        publicImported: { type: number }
        namedEligible: { type: number }
        publicEligible: { type: number }
        idempotent: { type: boolean }
        exclusions: { type: object, additionalProperties: { type: number } }
    OutreachLaneConfig:
      type: object
      required: [campaignKey, effectiveMonth, namedLimit, publicLimit, hotMarketLimit, hotLeadLimit]
      properties:
        campaignKey: { type: string }
        effectiveMonth: { type: string }
        namedLimit: { type: number }
        publicLimit: { type: number }
        hotMarketLimit: { type: number }
        hotLeadLimit: { type: number }
    OutreachOneTimeOverrideInput:
      type: object
      required: [campaignKey, requestedForDate, requestedLimit]
      properties:
        campaignKey: { type: string }
        requestedForDate: { type: string, format: date }
        requestedLimit: { type: number, minimum: 1 }
        reason: { type: string }
    OutreachOneTimeOverride:
      type: object
      required: [campaignKey, requestedForDate, requestedLimit, requestedBy, status, expiresAt]
      properties:
        campaignKey: { type: string }
        requestedForDate: { type: string, format: date }
        requestedLimit: { type: number }
        requestedBy: { type: string }
        status: { type: string, enum: [requested, blocked] }
        decisionEvidence: { type: string }
        expiresAt: { type: string, format: date-time }
...
    ErrorMessage:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    OutreachSystemConfig:
      type: object
      required: [version, policy, month, schedule]
      properties:
        version: { type: number }
        month: { type: string }
        policy: { type: object, additionalProperties: true }
        schedule: { type: object, additionalProperties: true }
...