# lib/api-zod/src/generated/types/inventoryImportReport.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InventoryImportReportExclusions } from './inventoryImportReportExclusions';

export interface InventoryImportReport {
  batchId: string;
  sourceRows: number;
  imported?: number;
  importedPendingQualification: number;
  excluded: number;
  namedImported: number;
  publicImported: number;
  namedEligible: number;
  publicEligible: number;
  idempotent?: boolean;
  exclusions: InventoryImportReportExclusions;
}