# lib/api-zod/src/generated/types/outreachOneTimeOverrideInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OutreachOneTimeOverrideInput {
  campaignKey: string;
  requestedForDate: Date;
  /** @minimum 1 */
  requestedLimit: number;
  reason?: string;
}